chore(sheet): set null cell util
