+++ packages/sheets-ui/src/commands/commands/auto-fill.command.ts
@@ -17,10 +17,9 @@
 import type { ICommand, IRange } from '@univerjs/core';
 import { CommandType, ICommandService, IUndoRedoService, IUniverInstanceService } from '@univerjs/core';
 import type { ISetRangeValuesMutationParams } from '@univerjs/sheets';
-import { getSheetCommandTarget, SetRangeValuesMutation, SetRangeValuesUndoMutationFactory, SetSelectionsOperation } from '@univerjs/sheets';
+import { generateNullCellValue, getSheetCommandTarget, SetRangeValuesMutation, SetRangeValuesUndoMutationFactory, SetSelectionsOperation } from '@univerjs/sheets';
 import type { IAccessor } from '@wendellhu/redi';
 
-import { generateNullCellValue } from '../../services/auto-fill/tools';
 import { IAutoFillService } from '../../services/auto-fill/auto-fill.service';
 
 export interface IAutoFillCommandParams {
