+++ packages/sheets/src/index.ts
@@ -378,3 +378,4 @@ export { RangeProtectionPermissionEditPoint } from './services/permission/permis
 export { RangeProtectionPermissionViewPoint } from './services/permission/permission-point/range/view';
 
 export type { IAddWorksheetProtectionParams } from './commands/mutations/add-worksheet-protection.mutation';
+export { generateNullCellValue } from './basics/utils';
